feat(matchengine): update data types to match spec (uint64/uint32)
- Updated Order struct with proper fields: ID (uint64), UserID (uint32), Left, DealStock, etc.
- Updated OrderSide: SideAsk=1, SideBid=2
- Updated OrderStatus: Pending=1, Partial=2, Finished=3, Canceled=4
- Updated Balance struct with Available and Frozen fields
- Updated BalanceManager with Freeze/Unfreeze/Add/Sub methods
- Updated IDGenerator to use uint64
- Updated Trade struct with proper uint64/uint32 fields
- Updated all handler files to use new types
- Fixed integration tests to use new types

+++ services/matchengine/cmd/main.go
@@ -129,10 +129,10 @@ func loadBalanceFromDB(db *sql.DB, e *engine.Engine) error {
 		}
 
 		if balType == 0 {
-			e.SetBalance(userID, asset, balance, decimal.Zero)
+			e.SetBalance(uint32(userID), asset, balance, decimal.Zero)
 			loaded++
 		} else if balType == 1 {
-			e.SetBalance(userID, asset, balance, decimal.Zero)
+			e.SetBalance(uint32(userID), asset, balance, decimal.Zero)
 			loaded++
 		}
 	}
